# bloop-core/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Bloop!","version":"0.1.0"},"paths":{"/video/generate":{"post":{"tags":["SadTalker Vdeo"],"summary":"Generate Video","operationId":"generate_video_video_generate_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_generate_video_video_generate_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/video/status/{job_id}":{"get":{"tags":["SadTalker Vdeo"],"summary":"Get Video","operationId":"get_video_video_status__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/video/result/{job_id}":{"get":{"tags":["SadTalker Vdeo"],"summary":"Get Video","operationId":"get_video_video_result__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/qa/upload-doc":{"post":{"tags":["Document Ask Endpoints"],"summary":"Upload File","operationId":"upload_file_qa_upload_doc_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_qa_upload_doc_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/qa/ask":{"post":{"tags":["Document Ask Endpoints"],"summary":"Ask Question","operationId":"ask_question_qa_ask_post","parameters":[{"name":"question","in":"query","required":true,"schema":{"type":"string","title":"Question"}},{"name":"document_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"}},{"name":"video_enabled","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Video Enabled"}},{"name":"image_path","in":"query","required":false,"schema":{"type":"string","default":"D:/bloop/data/avatars/sir-isaac-newton.webp","title":"Image Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/qa/ask-from-image":{"post":{"tags":["Document Ask Endpoints"],"summary":"Ask From Image","operationId":"ask_from_image_qa_ask_from_image_post","parameters":[{"name":"document_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Document Id"}},{"name":"video_enabled","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Video Enabled"}},{"name":"image_path","in":"query","required":false,"schema":{"type":"string","default":"D:/bloop/data/avatars/sir-isaac-newton.webp","title":"Image Path"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ask_from_image_qa_ask_from_image_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tts/answer":{"post":{"tags":["TTS"],"summary":"Speak Answer","operationId":"speak_answer_tts_answer_post","parameters":[{"name":"text","in":"query","required":true,"schema":{"type":"string","title":"Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/flashcards/{document_id}":{"get":{"tags":["Flashcards"],"summary":"Get Flashcards","operationId":"get_flashcards_flashcards__document_id__get","parameters":[{"name":"docuemnt_id","in":"query","required":true,"schema":{"type":"string","title":"Docuemnt Id"}},{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/quiz/{document_id}":{"get":{"tags":["Quiz"],"summary":"Get Quiz","operationId":"get_quiz_quiz__document_id__get","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}},{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/teach-ai/start-session":{"post":{"tags":["Play | Teach AI"],"summary":"Teach Ai Start Session","description":"LLM initiates the Teach-AI conversation by asking\nthe first question based on the concept.","operationId":"teach_ai_start_session_play_teach_ai_start_session_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_teach_ai_start_session_play_teach_ai_start_session_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/teach-ai/evaluate-response":{"post":{"tags":["Play | Teach AI"],"summary":"Teach Ai Evaluate Response","description":"Evaluates the user's explanation (text or audio),\ngenerates feedback + follow-up question,\nand optionally returns TTS audio.","operationId":"teach_ai_evaluate_response_play_teach_ai_evaluate_response_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_teach_ai_evaluate_response_play_teach_ai_evaluate_response_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeachAIResponsePayload"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/find-mistake/generate":{"post":{"tags":["Play | Find the Mistake"],"summary":"Generate Mistake","description":"Generates a flawed artifact based on:\n- concept\n- level\n- mistake type (optional)","operationId":"generate_mistake_play_find_mistake_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMistakeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMistakeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/find-mistake/evaluate":{"post":{"tags":["Play | Find the Mistake"],"summary":"Evaluate Mistake","description":"Evaluates learner's correction against the stored flawed artifact.","operationId":"evaluate_mistake_play_find_mistake_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateMistakeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateMistakeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/complete-missing-link/generate":{"post":{"tags":["Play | Complete Missing Link"],"summary":"Generate Missing Link","description":"Generates a drag-and-drop 'complete the missing link' puzzle.","operationId":"generate_missing_link_play_complete_missing_link_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMissingLinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateMissingLinkResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/play/complete-missing-link/evaluate":{"post":{"tags":["Play | Complete Missing Link"],"summary":"Evaluate Missing Link","description":"Evaluates learner's filled slots against the stored solution.","operationId":"evaluate_missing_link_play_complete_missing_link_evaluate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateMissingLinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EvaluateMissingLinkResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health","operationId":"health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_ask_from_image_qa_ask_from_image_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_ask_from_image_qa_ask_from_image_post"},"Body_generate_video_video_generate_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"},"audio":{"type":"string","format":"binary","title":"Audio"}},"type":"object","required":["image","audio"],"title":"Body_generate_video_video_generate_post"},"Body_teach_ai_evaluate_response_play_teach_ai_evaluate_response_post":{"properties":{"session_id":{"type":"string","title":"Session Id"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"},"audio":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Audio"},"voice_mode":{"type":"boolean","title":"Voice Mode","default":false}},"type":"object","required":["session_id"],"title":"Body_teach_ai_evaluate_response_play_teach_ai_evaluate_response_post"},"Body_teach_ai_start_session_play_teach_ai_start_session_post":{"properties":{"concept_id":{"type":"string","title":"Concept Id"},"level":{"type":"string","title":"Level","default":"beginner"}},"type":"object","required":["concept_id"],"title":"Body_teach_ai_start_session_play_teach_ai_start_session_post"},"Body_upload_file_qa_upload_doc_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_qa_upload_doc_post"},"EvaluateMissingLinkRequest":{"properties":{"session_id":{"type":"string","title":"Session Id"},"answers":{"additionalProperties":{"type":"string"},"type":"object","title":"Answers"}},"type":"object","required":["session_id","answers"],"title":"EvaluateMissingLinkRequest"},"EvaluateMissingLinkResponse":{"properties":{"correct":{"type":"boolean","title":"Correct"},"feedback":{"items":{"type":"string"},"type":"array","title":"Feedback"},"follow_up_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Follow Up Question"}},"type":"object","required":["correct","feedback","follow_up_question"],"title":"EvaluateMissingLinkResponse"},"EvaluateMistakeRequest":{"properties":{"session_id":{"type":"string","title":"Session Id"},"learner_fix":{"type":"string","title":"Learner Fix"},"explanation":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Explanation"}},"type":"object","required":["session_id","learner_fix"],"title":"EvaluateMistakeRequest"},"EvaluateMistakeResponse":{"properties":{"correct":{"type":"boolean","title":"Correct"},"feedback":{"items":{"type":"string"},"type":"array","title":"Feedback"},"follow_up_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Follow Up Question"}},"type":"object","required":["correct","feedback","follow_up_question"],"title":"EvaluateMistakeResponse"},"GenerateMissingLinkRequest":{"properties":{"concept_id":{"type":"string","title":"Concept Id"},"level":{"type":"string","title":"Level"},"category":{"type":"string","title":"Category"}},"type":"object","required":["concept_id","level","category"],"title":"GenerateMissingLinkRequest"},"GenerateMissingLinkResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"structure":{"items":{"$ref":"#/components/schemas/Slot"},"type":"array","title":"Structure"},"options":{"items":{"$ref":"#/components/schemas/Option"},"type":"array","title":"Options"}},"type":"object","required":["session_id","structure","options"],"title":"GenerateMissingLinkResponse"},"GenerateMistakeRequest":{"properties":{"concept_id":{"type":"string","title":"Concept Id"},"level":{"type":"string","title":"Level"},"mistake_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mistake Type"}},"type":"object","required":["concept_id","level"],"title":"GenerateMistakeRequest"},"GenerateMistakeResponse":{"properties":{"session_id":{"type":"string","title":"Session Id"},"artifact_type":{"type":"string","title":"Artifact Type"},"content":{"type":"string","title":"Content"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","required":["session_id","artifact_type","content"],"title":"GenerateMistakeResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Option":{"properties":{"option_id":{"type":"string","title":"Option Id"},"text":{"type":"string","title":"Text"}},"type":"object","required":["option_id","text"],"title":"Option"},"Slot":{"properties":{"slot_id":{"type":"string","title":"Slot Id"},"text":{"type":"string","title":"Text"}},"type":"object","required":["slot_id","text"],"title":"Slot"},"TeachAIEvaluationResponse":{"properties":{"scores":{"additionalProperties":true,"type":"object","title":"Scores"},"feedback":{"items":{"type":"string"},"type":"array","title":"Feedback"},"follow_up_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Follow Up Question"},"passed":{"type":"boolean","title":"Passed"}},"type":"object","required":["scores","feedback","follow_up_question","passed"],"title":"TeachAIEvaluationResponse"},"TeachAIResponsePayload":{"properties":{"evaluation":{"$ref":"#/components/schemas/TeachAIEvaluationResponse"},"assistant_text":{"type":"string","title":"Assistant Text"},"audio_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Audio Path"}},"type":"object","required":["evaluation","assistant_text","audio_path"],"title":"TeachAIResponsePayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}